+++ CHANGELOG.md
@@ -2,12 +2,17 @@
 
 ##
 
+
 ### Added
 - Tests for SSX experiment functions
+
 ### Changed
 - Refactoring of I24_Eiger into SSX_Eiger in order to have just one common tool to more beamlines for ssx collections.
 - Tidying up of SSX_Tristan.
 - Removed ssx from I19-2 cli.
+
+### Fixed
+- In CopyTristanNexus, `single_image_nexus` now takes number of bins for a static pump-probe experiment as input instead of a flag.
 
 ## 0.6.19